to find area of traingle with base and height with even float values

+++ src/Main.java
@@ -4,12 +4,12 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc=new Scanner(System.in);
 
-int B,H;
+float B,H;
         System.out.println("enter base");
-        B=sc.nextInt();
+        B=sc.nextFloat();
         System.out.println("enter height");
-        H=sc.nextInt();
-        int ans=(B*H)/2;
+        H=sc.nextFloat();
+        float ans=(B*H)/2;
 
 
         System.out.println(ans);
